Another types change (hopefully last for NFS and ditool).
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_softfloat@1693 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/ditool/im.h
@@ -102,8 +102,8 @@ struct disk_label {
 
 struct im_t {
     FILE*             imf;
-    unsigned          diskOffset;
-    unsigned          blockSize;
+    size_t            diskOffset;
+    size_t            blockSize;
     bool              rawOptical;
 
     struct disk_label dl;
